requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
docopt==0.6.2
exceptiongroup==1.0.4
Flask==2.2.2
h11==0.14.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.74
MarkupSafe==2.1.1
MouseInfo==0.1.3
outcome==1.2.0
packaging==22.0
Pillow==9.3.0
pipwin==0.5.2
PyAudio @ file:///C:/Users/daso/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==0.1.28
pySmartDL==1.3.4
PySocks==1.7.1
pyttsx3==2.90
pytweening==1.0.4
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
pywin32==305
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
speedtest-cli==2.1.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
Werkzeug==2.2.2
wikipedia==1.4.0
wsproto==1.2.0
zipp==3.11.0
